+++ releases/ChangeLog.md
@@ -1,7 +1,16 @@
-## 3.2.2 - Jan ??, 2021
-- Bugfix: Disable more path derivation checks to "Skip Checks" feature for multisig
-  compatibility. Addresses error shown when working previous Spectre multisig wallets
-  (`multisig.py: 891`)
+## 3.2.2 - Jan 14, 2021
+
+- Major Address Explorer enhancements! Thanks go to [@switck](https://twitter.com/switck)
+  for this major bump.
+    - View sub-accounts now, just enter the account number.
+    - Enter any custom derivation path, by entering numbers directly.
+    - Multisig wallet support! (Caveat: addresses are for verification purposes
+      and never for direct use as deposit, so they are partially redacted)
+    - Warning screen can be suppressed after reading first time (press 6)
+    - Export of addresses now named "addresses.csv" not "...txt"
+- Bugfix: Disable more path derivation checks for "Skip Checks" in
+  multisig compatibility feature. Handles error shown when working
+  with previously-imported Spectre multisig wallets (`multisig.py: 891`).
 
 ## 3.2.1 - Jan 8, 2021